requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-logging>=3.0.0